PWN-6964 - Fix notification permission event value
