+++ app/src/main/java/org/p2p/wallet/auth/ui/onboarding/root/OnboardingRootContract.kt
@@ -11,7 +11,5 @@ interface OnboardingRootContract {
         fun navigateToCreatePin()
     }
 
-    interface Presenter : MvpPresenter<View> {
-        fun logNotificationPermissionGranted(isGranted: Boolean)
-    }
+    interface Presenter : MvpPresenter<View>
 }
