+++ app/src/main/java/org/p2p/wallet/auth/ui/onboarding/root/OnboardingRootPresenter.kt
@@ -45,8 +45,4 @@ class OnboardingRootPresenter(
     private fun userLeftOnPinCreation(): Boolean {
         return tokenKeyProvider.keyPair.isNotEmpty() && !authInteractor.isAuthorized()
     }
-
-    override fun logNotificationPermissionGranted(isGranted: Boolean) {
-        onboardingAnalytics.setUserGrantedNotificationPermissions(isGranted = isGranted)
-    }
 }
