+++ app/src/main/java/org/p2p/wallet/splash/SplashContract.kt
@@ -10,5 +10,7 @@ interface SplashContract {
         fun navigateToMain()
     }
 
-    interface Presenter : MvpPresenter<View>
+    interface Presenter : MvpPresenter<View> {
+        fun logNotificationPermissionGranted(isGranted: Boolean)
+    }
 }
